Creating context for auth status
