+++ app/src/components/Dashboard.tsx
@@ -1,12 +1,5 @@
-import React, { useEffect, useState } from "react";
+import React from "react";
 import { gql, useQuery } from "@apollo/client";
-
-interface IUser {
-  identityProvider: string;
-  userId: string;
-  userDetails: string;
-  userRoles: string[];
-}
 
 const authorsQuery = gql`
   query {
@@ -34,25 +27,14 @@ const GraphQLData = () => {
 };
 
 const Dashboard = () => {
-  const [user, setUser] = useState<IUser | null>(null);
-  useEffect(() => {
-    const fetchAuth = async () => {
-      const response = await fetch("/.auth/me");
-      const payload = await response.json();
-      const { clientPrincipal } = payload;
-      setUser(clientPrincipal);
-    };
-    fetchAuth();
-  }, []);
-
   return (
     <div>
-      <h3>Homepage</h3>
+      {/* <h3>Homepage</h3>
       {user?.userId ? (
         <h5>Authenticated: {user.userId}</h5>
       ) : (
         <h5>Unauthenticated</h5>
-      )}
+      )} */}
       <h4>GraphQL Data</h4>
       <GraphQLData />
     </div>
