+++ app/src/components/Homepage.tsx
@@ -1,7 +1,18 @@
 import React from "react";
+import { useAuth } from "../hooks/useAuth";
 
 const Homepage = () => {
-  return <h5>Homepage</h5>;
+  const { user } = useAuth();
+  return (
+    <div>
+      <h5>Homepage</h5>
+      <p>
+        {user?.userDetails
+          ? `Authenticated: ${user.userDetails}`
+          : "Not Authenticated!"}
+      </p>
+    </div>
+  );
 };
 
 export default Homepage;
