+++ app/src/components/PrivateRoute.tsx
@@ -1,0 +1,35 @@
+import React from "react";
+import { Route, Redirect } from "react-router-dom";
+
+// function PrivateRoute({ children, ...rest }) {
+//   return (
+//     <Route
+//       {...rest}
+//       render={({ location }) =>
+//         fakeAuth.isAuthenticated ? (
+//           children
+//         ) : (
+//           <Redirect
+//             to={{
+//               pathname: "/login",
+//               state: { from: location },
+//             }}
+//           />
+//         )
+//       }
+//     />
+//   );
+// }
+
+// const PrivateRoute = ({ component: Component, ...rest }) => {
+//   return (
+//     <Route
+//       {...rest}
+//       render={(props) =>
+//         authenticated ? <Component {...props} /> : <Redirect to="/login" />
+//       }
+//     />
+//   );
+// };
+
+// return PrivateRoute;
